changed hierarchy of our system + created a server.js file and moved configuration from index to it
