--- loginAndForm/market.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <link href="../favicon.ico" rel="icon" type="image/x-icon">
-    <script src="https://js.stripe.com/v3/"></script>
-    <script
-            crossorigin="anonymous"
-            integrity="sha256-3zlB5s2uwoUzrXK3BT7AX3FyvojsraNFxCc2vC/7pNI="
-            src="https://code.jquery.com/jquery-3.6.1.js"></script>
-    <script src="https://checkout.stripe.com/checkout.js"></script>
-    <script src="https://code.jquery.com/jquery-3.5.1.min.js"></script>
-
-    <script src="../js/market.js" type="module"></script>
-    <link href="../css/style.css" rel="stylesheet">
-    <meta charset="UTF-8">
-    <title>Billing Plans</title>
-    <link crossorigin="anonymous" href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta3/dist/css/bootstrap.min.css"
-          integrity="sha384-eOJMYsd53ii+scO/bJGFsiCZc+5NDVN2yr8+0RDqr0Ql0h+rP48ckxlpbzKgwra6" rel="stylesheet">
-</head>
-<body>
-<div class="login-box" id="form-box">
-    <header>
-        <h1>Welcome, these are our plans:</h1>
-    </header>
-    <main>
-        <div class="container">
-        </div>
-        <div></div>
-    </main>
-</div>
-</body>
-</html>
